backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Carlos A Palacios R",email = "carlosargelio0104@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "django (>=5.2.8,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "django-configurations (>=2.5.1,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "drf-spectacular (>=0.29.0,<0.30.0)"
]

[tool.poetry]
packages = [{include = "backend", from = "src"}]


[tool.poetry.group.dev.dependencies]
mypy = "^1.18.2"
ruff = "^0.14.4"
pre-commit = "^4.4.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
files = "**/*.py"

ignore_missing_imports = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
# strict = true  # Recomendado para proyectos maduros

[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    "**/.git",
    "__pycache__",
]

select = ["E", "F", "I", "N", "B", "T", "W"]

line-length = 88

[tool.ruff.lint]
ignore = ["E501"] # Ignorar E501 (línea demasiado larga) ya que line-length ya lo maneja.

[tool.ruff.format]
# Configuración del formateador (generalmente usa las configuraciones del linter)
